Move POST endpoint to separate GET request. change to Query params
